2006-04-29  Peter Dennis Bartok  <pbartok@novell.com>
	* XplatUIX11.cs (SetWindowTransparency): Casting opacity to int before
	  casting to IntPtr to avoid 64bit overflow errors


svn path=/trunk/mcs/; revision=60091

+++ mcs/class/Managed.Windows.Forms/System.Windows.Forms/XplatUIX11.cs
@@ -4234,7 +4234,7 @@ namespace System.Windows.Forms {
 			}
 
 			hwnd.opacity = (uint)(0xffffffff * transparency);
-			opacity = (IntPtr)hwnd.opacity;
+			opacity = (IntPtr)((int)hwnd.opacity);
 
 			if (hwnd.reparented) {
 				XChangeProperty(DisplayHandle, XGetParent(hwnd.whole_window), NetAtoms[(int)NA._NET_WM_WINDOW_OPACITY], (IntPtr)Atom.XA_CARDINAL, 32, PropertyMode.Replace, ref opacity, 1);
